util_unlink_path() -> util_clear_directory() + added PRE_CLEAR_RUNPATH config alternative + ecl_smspec changes
git-svn-id: http://ert.nr.no/repos/ert@2341 ecb61f28-d21e-0410-a4b6-a828cfe0ca3a

+++ devel/libutil/src/util_path.c
@@ -15,19 +15,6 @@ bool util_is_abs_path(const char * path) {
 }
 
 
-static int util_get_path_length(const char * file) {
-  if (util_is_directory(file)) 
-    return strlen(file);
-  else {
-    char * last_slash = strrchr(file , UTIL_PATH_SEP_CHAR);
-    if (last_slash == NULL)
-      return 0;
-    else 
-      return last_slash - file;
-  }
-}
-
-
 
 static bool util_make_path2__(const char *path , mode_t mode) {
   /* It already exists as a directory - we just return true. */
@@ -252,55 +239,54 @@ char * util_realloc_filename(char * filename , const char * path , const char * 
 /**
    Only removes the last component in path.
 */
-void static util_unlink_path_static( const char *path ) {
+void static util_clear_directory__( const char *path , bool strict_uid , bool unlink_root) {
   if (util_is_directory(path)) {
-    const uid_t uid = getuid();
-    DIR  *dirH;
-    struct dirent *dentry;
-    dirH = opendir( path );  
+    DIR  *dirH = opendir( path );
 
-    while ( (dentry = readdir(dirH)) != NULL) {
-      struct stat buffer;
-      mode_t mode;
-      const char * entry_name = dentry->d_name;
-      if ((strcmp(entry_name , ".") != 0) && (strcmp(entry_name , "..") != 0)) {
-	char * full_path = util_alloc_filename(path , entry_name , NULL);
+    if (dirH != NULL) {
+      const uid_t uid = getuid();
+      struct dirent *dentry;
+      
+      while ( (dentry = readdir(dirH)) != NULL) {
+        struct stat buffer;
+        mode_t mode;
+        const char * entry_name = dentry->d_name;
+        if ((strcmp(entry_name , ".") != 0) && (strcmp(entry_name , "..") != 0)) {
+          char * full_path = util_alloc_filename(path , entry_name , NULL);
 
-	if (lstat(full_path , &buffer) != 0) {
-	  fprintf(stderr,"%s: failed to stat: %s entry not removed.\n",__func__ , full_path);
-	} else {
-	  mode = buffer.st_mode;
+          if (lstat(full_path , &buffer) == 0) {
+            mode = buffer.st_mode;
 	  
-	  if (S_ISDIR(mode)) 
-            /*
-              Recursively descending into sub directory. 
-            */
-	    util_unlink_path_static(full_path);
-          else if (S_ISLNK(mode)) 
-	    /*
-	      Symbolic links are unconditionally removed.
-	    */
-            unlink(full_path);
-	  else if (S_ISREG(mode)) {
-            /* 
-               It is a regular file - we remove it (if we own it!).
-            */
-	    if (buffer.st_uid == uid) {
-              int unlink_return = unlink(full_path);
-	      if (unlink_return != 0) {
-                /* Unlink failed - we don't give a shit. */
-              }
-	    } else 
-	      fprintf(stderr,"Warning mismatch in uid of calling process and owner for:%s - entry *not* removed \n",full_path);
-
-	  }
-	}
-	free(full_path);
+            if (S_ISDIR(mode)) 
+              /*
+                Recursively descending into sub directory. 
+              */
+              util_clear_directory__(full_path , strict_uid , true);
+            else if (S_ISLNK(mode)) 
+              /*
+                Symbolic links are unconditionally removed.
+              */
+              unlink(full_path);
+            else if (S_ISREG(mode)) {
+              /* 
+                 It is a regular file - we remove it (if we own it!).
+              */
+              if ((!strict_uid) || (buffer.st_uid == uid)) {
+                int unlink_return = unlink(full_path);
+                if (unlink_return != 0) {
+                  /* Unlink failed - we don't give a shit. */
+                }
+              } 
+            }
+          }
+          free(full_path);
+        }
       }
     }
-    
     closedir(dirH);
-    {
+
+    /* Finish with clearing the root directory */
+    if (unlink_root) {
       int rmdir_return = rmdir(path);
       if (rmdir_return != 0) {
         /* Unlink failed - we don't give a shit. */
@@ -310,9 +296,24 @@ void static util_unlink_path_static( const char *path ) {
 }
 
 
+/**
+   This function will clear away all the contents (including
+   subdirectories) in the directory @path.
 
-void util_unlink_path(const char *path) {
-  util_unlink_path_static( path );
+   If the parameter @strict_uid is set to true, the function will only
+   attempt to remove entries where the calling uid is also the owner
+   of the entry. 
+
+   If the parameter @unlink_root is true the directory @path will also
+   be removed, otherwise it will be left as an empty directory.
+
+   The The function will just go about deleting as much as it can; errors
+   are not signalled in any way!
+*/
+
+
+void util_clear_directory(const char * path , bool strict_uid , bool unlink_root) {
+  util_clear_directory__( path , strict_uid , unlink_root );
 }
